# docs/swagger.yaml
definitions:
  dto.Geometry:
    properties:
      location:
        $ref: '#/definitions/dto.Location'
    type: object
  dto.GetEvent:
    properties:
      end_time:
        type: string
      id:
        type: string
      place_id:
        type: string
      start_time:
        type: string
      trip_id:
        type: string
    type: object
  dto.GetUser:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: integer
      login:
        type: string
      password:
        type: string
    type: object
  dto.GooglePlace:
    properties:
      formatted_address:
        type: string
      geometry:
        $ref: '#/definitions/dto.Geometry'
      name:
        type: string
      photos:
        items:
          $ref: '#/definitions/dto.Photo'
        type: array
      place_id:
        type: string
      rating:
        type: number
      types:
        items:
          type: string
        type: array
      vicinity:
        type: string
    type: object
  dto.Location:
    properties:
      lat:
        type: number
      lng:
        type: number
    type: object
  dto.Photo:
    properties:
      photo_reference:
        type: string
    type: object
  dto.TripResponse:
    properties:
      area:
        $ref: '#/definitions/dto.GooglePlace'
      area_id:
        type: string
      end_time:
        type: string
      events:
        items:
          $ref: '#/definitions/dto.GetEvent'
        type: array
      id:
        type: string
      name:
        type: string
      places:
        items:
          $ref: '#/definitions/dto.GooglePlace'
        type: array
      start_time:
        type: string
      users:
        items:
          $ref: '#/definitions/dto.GetUser'
        type: array
    type: object
  handler.AddPlaceToTripRequest:
    properties:
      place_id:
        type: string
      trip_id:
        type: string
    required:
    - place_id
    - trip_id
    type: object
...
  handler.UpdateUserRequest:
    properties:
      email:
        type: string
      id:
        type: integer
      login:
        type: string
      password:
        type: string
    required:
    - email
    - id
    - login
    - password
    type: object
  model.Event:
    properties:
      endTime:
        type: string
      id:
        type: string
      name:
        type: string
      placeID:
        type: string
      startTime:
        type: string
      tripID:
        type: string
    type: object
  model.Geometry:
    properties:
      location:
        $ref: '#/definitions/model.Location'
    type: object
  model.GooglePlace:
    properties:
      formatted_address:
        type: string
      geometry:
        $ref: '#/definitions/model.Geometry'
      name:
        type: string
      photos:
        items:
          $ref: '#/definitions/model.Photo'
        type: array
      place_id:
        type: string
      rating:
        type: number
      types:
        items:
          type: string
        type: array
      vicinity:
        type: string
    type: object
  model.Location:
    properties:
      lat:
        type: number
      lng:
        type: number
    type: object
  model.Photo:
    properties:
      photo_reference:
        type: string
    type: object
  model.Place:
    properties:
      closing:
        description: |-
          Name        string  `json:"name"`
          Photo       string  `json:"photo"`
          Rating      float32 `json:"rating"`
        type: string
      googlePlace:
        $ref: '#/definitions/model.GooglePlace'
      id:
        type: string
      opening:
        type: string
      trips:
        items:
          $ref: '#/definitions/model.Trip'
        type: array
    type: object
  model.Trip:
    properties:
      area:
        $ref: '#/definitions/model.Place'
      area_id:
        type: string
      end_time:
        type: string
      events:
        items:
          $ref: '#/definitions/model.Event'
        type: array
      id:
        type: string
      name:
        type: string
      places:
        items:
          $ref: '#/definitions/model.Place'
        type: array
      start_time:
        type: string
      users:
        items:
          $ref: '#/definitions/model.User'
        type: array
    type: object
  model.User:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: integer
      image_url:
        type: string
      login:
        type: string
      password:
        type: string
    type: object
info:
  contact: {}
paths:
...
  /api/v1/place:
    get:
      description: Find places by name or location
      parameters:
      - description: Name to search places by
        in: query
        name: name
        required: true
        type: string
      - description: Type of place
        in: query
        name: type
        type: string
      - description: Latitude for location-based search
        in: query
        name: lat
        type: string
      - description: Longitude for location-based search
        in: query
        name: lng
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of found places
          schema:
            additionalProperties:
              items:
                $ref: '#/definitions/model.Place'
              type: array
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get places
      tags:
      - place
  /api/v1/place/find:
    get:
      description: Find places by searchString
      parameters:
      - description: Search string to search places
        in: query
        name: searchString
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of found places
          schema:
            additionalProperties:
              items:
                $ref: '#/definitions/dto.GooglePlace'
              type: array
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Find places
      tags:
      - place
  /api/v1/place/photo:
    get:
      description: Get a photo of a place by photo reference
      parameters:
      - description: Photo reference ID
        in: query
        name: reference
        required: true
        type: string
      produces:
      - image/jpeg
      responses:
        "200":
          description: Binary image data of the place photo
          schema:
            type: file
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get place photo
      tags:
      - place
...
  /api/v1/trip/{trip_id}/place/{place_id}:
    delete:
      consumes:
      - application/json
      description: Delete place from a specific trip by their IDs
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.TripResponse'
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete place from trip
      tags:
      - place
...
  /api/v1/trip/place:
    post:
      consumes:
      - application/json
      description: Add a place to a specific trip by their IDs
      parameters:
      - description: JSON containing trip and place IDs
        in: body
        name: trip-place
        required: true
        schema:
          $ref: '#/definitions/handler.AddPlaceToTripRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.TripResponse'
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Add place to trip
      tags:
      - place
  /api/v1/user:
    put:
      consumes:
      - application/json
      description: Updates the details of an existing user.
      parameters:
      - description: User data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/handler.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            properties:
              err:
                type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            properties:
              err:
                type: string
            type: object
        "404":
          description: Not Found
          schema:
            properties:
              err:
                type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            properties:
              err:
                type: string
            type: object
      summary: Update user details
      tags:
      - user
  /api/v1/user/{user_id}:
    get:
      description: Retrieves a user by their ID.
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.User'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get user by ID
      tags:
      - user
swagger: "2.0"